modified rating filter widget

+++ parka/lib/pages/filter/components/star_rating_widget.dart
@@ -2,7 +2,7 @@ import 'package:ParkA/styles/parka_colors.dart';
 import "package:flutter/material.dart";
 
 class StarRatingFilter extends StatelessWidget {
-  final int rating;
+  final double rating;
   final Function rateHandler;
 
   const StarRatingFilter({
